docs: Add docs and done some reformatting
