+++ hope/src/main/java/io/github/madethoughts/hope/network/packets/serverbound/handshake/Handshake.java
@@ -19,7 +19,7 @@
 package io.github.madethoughts.hope.network.packets.serverbound.handshake;
 
 import io.github.madethoughts.hope.network.State;
-import io.github.madethoughts.hope.network.packets.Deserializer;
+import io.github.madethoughts.hope.network.packets.serverbound.Deserializer;
 import io.github.madethoughts.hope.network.packets.serverbound.ServerboundPacket;
 
 public record Handshake(
