+++ hope/src/main/java/io/github/madethoughts/hope/network/packets/serverbound/login/LoginStart.java
@@ -18,7 +18,7 @@
 
 package io.github.madethoughts.hope.network.packets.serverbound.login;
 
-import io.github.madethoughts.hope.network.packets.Deserializer;
+import io.github.madethoughts.hope.network.packets.serverbound.Deserializer;
 import io.github.madethoughts.hope.network.packets.serverbound.ServerboundPacket;
 
 import java.util.UUID;
